+++ engine/src/service/time.hpp
@@ -1,17 +1,18 @@
 #pragma once
 
+#include "common.hpp"
 #include "service.hpp"
 
 namespace smeg {
 	
 	struct Time: public Service {
-		static constexpr float MS_TO_SEC = 0.001f;
-		static constexpr float SEC_TO_MS = 1000.0f;
+		static constexpr double MS_TO_SEC = 0.001f;
+		static constexpr double SEC_TO_MS = 1000.0f;
 
-		long ticks;
-		int fps;
+		u64 ticks;
+		double fps;
 
-		int targetFrameRate = 60;
+		i32 targetFrameRate = 60;
 
 		double current;
 		double lastFrame;
